+++ src/assets/script.js
@@ -4,10 +4,21 @@ $(document).ready(function () {
   $("#demo-rating").load("articles/rating.html", () => {
     initRating();
   });
-  $("#demo-uxstars").load("articles/uxstars.html", function (res, status, xhr) {
-    Prism.highlightAll();
+  $("#demo-uxstars").load("articles/uxstars.html");
+  $("#demo-tabs").load("articles/tabs.html", () => {
+    var demoTabs = [
+      { name: 'Confiture de cerise', id: 'cerise' },
+      { name: 'Frites belges', id: 'frites' },
+      { name: 'Cookies', id: 'cookies' },
+      { name: 'Limonade', id: 'limonade' }
+    ];
+    document.querySelector("pea11y-tabs").tabsNames = demoTabs;
   });
-  $("#demo-snippet").load("articles/snippet.html", function (res, status, xhr) {
-    initSnippet();
-  });
+  $("#demo-snippet").load("articles/snippet.html");
 });
+
+
+// UXSTARS
+function resetUxstars(idElem) {
+  document.getElementById(idElem).setAttribute("value", 0);
+}