+++ src/components/snippet/changelog.md
@@ -1,5 +1,10 @@
 # PeA11y snippet
 
+## [1.0.0] - 2018-08-14
+### Added
+- Propriete *aria-label* pour l'element avec le role *tablist* pour etre conforme au pattern
+- Boucle sur les onglets quand arrive a la fin ou au debut
+
 ## [1.0.0] - 2018-07-30
 ### Added
 - Fetch les fichiers grace a l'uri renseignee